// src/models/Gastos_compras.model.ts
export class GastoCompra {
    id: number | undefined;
    tipoGasto: string | undefined;
    concepto: string | undefined;
    fecha: Date | undefined;
    importe: number | undefined;
    recibio: string | undefined;
    autoriza: string | undefined;
    empleado: string | undefined;
    sucursal: string | undefined;
  }
  